requirements.txt
alembic==0.8.7
arrow==0.8.0
barcodenumber==0.2.1
blinker==1.4
boto==2.39.0
boto3==1.4.7
botocore==1.7.28
bottlenose==1.1.3
cachetools==2.0.1
certifi==2017.7.27.1
cffi==1.5.2
chardet==3.0.4
click==6.6
cryptography==1.3.1
dateparser==0.3.4
decorator==4.0.10
dill==0.2.7.1
docopt==0.4.0
docutils==0.14
enum34==1.1.6
Flask==0.10.1
Flask-Login==0.3.2
Flask-Mail==0.9.1
Flask-Migrate==1.8.0
Flask-Script==2.0.5
Flask-SQLAlchemy==2.1
future==0.16.0
futures==3.1.1
gax-google-logging-v2==0.8.3
gax-google-pubsub-v1==0.8.3
google-auth==1.1.1
google-cloud-core==0.27.1
google-cloud-storage==1.4.0
google-gax==0.15.15
google-resumable-media==0.3.0
googleapis-common-protos==1.5.3
grpc-google-logging-v2==0.8.1
grpc-google-pubsub-v1==0.8.1
gunicorn==19.4.5
idna==2.6
ipaddress==1.0.16
itsdangerous==0.24
jdatetime==1.7.4
Jinja2==2.8
jmespath==0.9.3
jws==0.1.3
mailchimp==2.0.9
Mako==1.0.4
MarkupSafe==0.23
memory-profiler==0.47
ndg-httpsclient==0.4.0
nltk==3.2.5
oauthlib==1.0.3
olefile==0.44
ply==3.8
praw==4.1.0
prawcore==0.5.0
protobuf==3.4.0
psutil==5.4.0
psycopg2==2.6.1
pyasn1==0.3.7
pyasn1-modules==0.1.5
pycparser==2.14
pyOpenSSL==16.0.0
pypng==0.0.18
PyQRCode==1.2.1
python-amazon-mws==0.6
python-dateutil==2.4.2
python-editor==1.0.1
python-firebase==1.2
pytz==2016.3
PyYAML==3.11
qrcode==5.3
redis==2.10.5
regex==2016.3.26
requests==2.18.4
requests-oauthlib==0.6.1
rq==0.6.0
rq-dashboard==0.3.7
rsa==3.4.2
ruamel.yaml==0.15.34
s3transfer==0.1.11
semantics3==0.3.4
six==1.11.0
SQLAlchemy==1.0.14
statistics==1.0.3.5
stripe==1.32.0
tzlocal==1.4
umalqurra==0.2
update-checker==0.15
urllib3==1.22
urlnorm==1.1.3
Werkzeug==0.11.5
